refactor(src): Make extensions use mp4uploadextractor lib
