+++ src/ar/animeblkom/build.gradle
@@ -10,6 +10,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mp4upload-extractor'))
     implementation(project(":lib-okru-extractor"))
     implementation("dev.datlag.jsunpacker:jsunpacker:1.0.1")
 }
