+++ src/ar/animetitans/build.gradle
@@ -10,6 +10,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mp4upload-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
 }
