+++ src/en/kissanime/build.gradle
@@ -11,6 +11,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mp4upload-extractor'))
     implementation(project(':lib-yourupload-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
 }
