+++ src/es/monoschinos/src/eu/kanade/tachiyomi/animeextension/es/monoschinos/MonosChinos.kt
@@ -5,7 +5,6 @@ import android.content.SharedPreferences
 import android.util.Base64
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
-import eu.kanade.tachiyomi.animeextension.es.monoschinos.extractors.Mp4uploadExtractor
 import eu.kanade.tachiyomi.animeextension.es.monoschinos.extractors.SolidFilesExtractor
 import eu.kanade.tachiyomi.animeextension.es.monoschinos.extractors.UploadExtractor
 import eu.kanade.tachiyomi.animesource.ConfigurableAnimeSource
@@ -15,6 +14,7 @@ import eu.kanade.tachiyomi.animesource.model.SAnime
 import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.ParsedAnimeHttpSource
+import eu.kanade.tachiyomi.lib.mp4uploadextractor.Mp4uploadExtractor
 import eu.kanade.tachiyomi.lib.okruextractor.OkruExtractor
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.util.asJsoup
@@ -94,8 +94,8 @@ class MonosChinos : ConfigurableAnimeSource, ParsedAnimeHttpSource() {
                     if (video != null) videoList.add(video)
                 }
                 url.contains("mp4upload") -> {
-                    val videoHeaders = headersBuilder().add("Referer", "https://mp4upload.com/").build()
-                    videoList.add(Mp4uploadExtractor().getVideoFromUrl(url, videoHeaders))
+                    val videos = Mp4uploadExtractor(client).videosFromUrl(url, headers)
+                    videoList.addAll(videos)
                 }
             }
         }
