+++ src/tr/turkanime/build.gradle
@@ -5,6 +5,7 @@ plugins {
 }
 
 dependencies {
+    implementation(project(':lib-mp4upload-extractor'))
     implementation(project(":lib-cryptoaes"))
     implementation(project(":lib-synchrony"))
     implementation(project(":lib-voe-extractor"))
